Move sideEffects from TradeForm to hook in Trade
